backend_django/requirements.txt
Django>=5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
python-dotenv>=1.0.0
PyJWT>=2.8.0
scikit-learn>=1.3.2
nltk>=3.8.1
spacy>=3.7.2
textblob>=0.17.1
numpy>=1.26.2
pandas>=2.1.4
joblib>=1.3.2
gunicorn>=21.2.0
bcrypt>=4.0.1
